Rewrote IfExecutionPhaseIsThen, refractor ExecuteIf
